orderLabel changed.

+++ Model/Abstract Classes/OrderAC.cs
@@ -10,12 +10,12 @@ namespace Kadeka.Model
     abstract class OrderAC<T>
     {
         int ID;
-        int TotalPrice;
+        float TotalPrice;
         string OrderNote;
         DateTime OrderDate;
         List<T> ProductList;
 
-        public OrderAC(int ID, int totalPrice, string orderNote)
+        public OrderAC(int ID, float totalPrice, string orderNote)
         {
             this.ID = ID;
             TotalPrice = totalPrice;
@@ -25,13 +25,13 @@ namespace Kadeka.Model
         }
 
         public int getID() { return ID; }
-        public int getTotalPrice() { return TotalPrice; }
+        public float getTotalPrice() { return TotalPrice; }
         public string getOrderNote() { return OrderNote; }
         public DateTime getOrderDate() { return OrderDate; }
         public List<T> getProductList() { return ProductList; }
 
         public void setID(int ID) { this.ID = ID; }
-        public void setTotalPrice(int totalPrice) { TotalPrice = totalPrice; }
+        public void setTotalPrice(float totalPrice) { TotalPrice = totalPrice; }
         public void setOrderNote(String note) { OrderNote = note; }
 
     }
